Added guard clause for AppId not defined in queryString and renamed load-tracker.ts
