+++ src/apply-tag/applyTag.ts
@@ -1,4 +1,4 @@
-import loadTracker from "./helpers/snowplow-events/load-tracker";
+import loadTracker from "./helpers/snowplow-events/instantiate-tracker";
 import signUp from "./helpers/snowplow-events/sign-up";
 import pageview from "./helpers/snowplow-events/pageview";
 import { chooseCart, chooseImpression } from "./helpers/dynamic-import";
@@ -6,6 +6,10 @@ import { Impressions, SignUp, Transactions } from "../shared/types";
 
 const applyTag = async (context) => {
   const isTrackerInitialized: Boolean = loadTracker(context as Transactions);
+
+  if(!context.appId) {
+    throw new Error("appId is required");
+  };
 
   if (isTrackerInitialized) {
     pageview(context as Transactions);
